Support for building under Linux (with SDL2).
