+++ NonEuclidean/Main.cpp
@@ -1,5 +1,11 @@
-#define _CRT_SECURE_NO_WARNINGS
+#if defined(_WIN32)
+	#define _CRT_SECURE_NO_WARNINGS
+#endif
 #include "Engine.h"
+
+#if defined(_WIN32)
+
+// --- Windows --------------------------------------------------------------
 
 int APIENTRY WinMain(HINSTANCE hCurrentInst, HINSTANCE hPreviousInst, LPSTR lpszCmdLine, int nCmdShow) {
   //Open console in debug mode
@@ -14,3 +20,14 @@ int APIENTRY WinMain(HINSTANCE hCurrentInst, HINSTANCE hPreviousInst, LPSTR lpsz
   Engine engine;
   return engine.Run();
 }
+
+#else
+
+// --- non-Windows ----------------------------------------------------------
+
+int main(int argc, char** argv) {
+	Engine engine;
+	return engine.Run();
+}
+
+#endif
